requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0